adding some integrety logic to service & repo classes

+++ src/main/java/com/example/springtp/service/ParticipationService.java
@@ -58,7 +58,7 @@ public class ParticipationService {
         return participationRepository.findByQuiz(quizId);
     }
 
-    public int calculateScore(Long participationId) {
+    public float calculateScore(Long participationId) {
         return participationRepository.calculateScore(participationId);
     }
 }
